Moving DartTour and other files to tours directory
