+++ src/org/tournament/Tournament.java
@@ -4,6 +4,7 @@ import java.util.*;
 import org.drivers.MatchDriver;
 import org.match_engine.*;
 import org.match_engine.dartbot.*;
+import org.tours.PrizeMoney;
 
 public class Tournament {
     public String name;
