+++ src/org/tours/PrizeMoney.java
@@ -1,4 +1,4 @@
-package org.tournament;
+package org.tours;
